requirements.txt
# Core dependencies for news producer
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0

# For future components
pymongo==4.6.1
pyspark==3.5.0

# Optional: for testing Kafka
kafka-python-ng==2.2.0